+++ RELEASENOTES.md
@@ -52,6 +52,8 @@ This release includes the following changes since [1.6.1](#161-2025-04-14):
         scrubber bar around). The behavior of scrubbing mode can be customized
         with `setScrubbingModeParameters(..)` on `ExoPlayer` and
         `ExoPlayer.Builder`.
+    *   Add support for using the virtual device ID from the `Context` passed to
+        `ExoPlayer.Builder`.
     *   Allow customizing fractional seek tolerance in scrubbing mode.
     *   Increase codec operating rate in scrubbing mode.
     *   Fix bug where prepare errors in the content of `AdsMediaSource` may be